feat: 升级 Codex Radar 并收口领域模块
