+++ qq-maid-core/src/runtime/respond/command_render.rs
@@ -9,35 +9,35 @@ use super::common::CommandBody;
 
 /// 同时维护 Markdown 与纯文本缓冲区的轻量 builder。
 #[derive(Debug, Default)]
-pub(super) struct CommandRender {
+pub(crate) struct CommandRender {
     text_lines: Vec<String>,
     markdown_lines: Vec<String>,
 }
 
 impl CommandRender {
-    pub(super) fn new() -> Self {
+    pub(crate) fn new() -> Self {
         Self::default()
     }
 
-    pub(super) fn title(&mut self, text: &str) {
+    pub(crate) fn title(&mut self, text: &str) {
         self.push_pair(text.to_owned(), format!("# {}", escape_inline(text)));
     }
 
-    pub(super) fn subtitle(&mut self, text: &str) {
+    pub(crate) fn subtitle(&mut self, text: &str) {
         self.push_pair(text.to_owned(), format!("## {}", escape_inline(text)));
     }
 
-    pub(super) fn paragraph(&mut self, text: &str) {
+    pub(crate) fn paragraph(&mut self, text: &str) {
         let text = text.trim().to_owned();
         self.push_pair(text.clone(), escape_text(&text));
     }
 
-    pub(super) fn bullet(&mut self, text: &str) {
+    pub(crate) fn bullet(&mut self, text: &str) {
         let text = text.trim().to_owned();
         self.push_pair(format!("· {text}"), format!("- {}", escape_text(&text)));
     }
 
-    pub(super) fn blank(&mut self) {
+    pub(crate) fn blank(&mut self) {
         if self
             .text_lines
             .last()
@@ -48,12 +48,12 @@ impl CommandRender {
         }
     }
 
-    pub(super) fn push_pair(&mut self, text: String, markdown: String) {
+    pub(crate) fn push_pair(&mut self, text: String, markdown: String) {
         self.text_lines.push(text);
         self.markdown_lines.push(markdown);
     }
 
-    pub(super) fn build(self) -> CommandBody {
+    pub(crate) fn build(self) -> CommandBody {
         CommandBody::dual(self.text_lines.join("\n"), self.markdown_lines.join("\n"))
     }
 }
