+++ qq-maid-core/src/runtime/respond/tests/mod.rs
@@ -7,7 +7,6 @@ mod memory_regressions;
 mod memory_session;
 mod memory_tool;
 mod ops;
-mod radar;
 mod rss;
 mod search;
 mod session;
