+++ qq-maid-core/src/runtime/respond/tests/support/executors.rs
@@ -322,6 +322,8 @@ pub(crate) fn mock_radar_snapshot() -> RadarSnapshot {
             window_message: Some("社区反馈已完成重置，当前没有开启的速蹬窗口".to_owned()),
             prediction_level: Some("high".to_owned()),
             probability_24h: Some(0.36),
+            probability_48h: Some(0.52),
+            prediction_summary: Some("短线仍需关注额度异常反馈。".to_owned()),
             model_score: Some(60.0),
             model_status: Some("red".to_owned()),
             model_passed: Some(4),
@@ -345,6 +347,22 @@ pub(crate) fn mock_radar_snapshot() -> RadarSnapshot {
             ],
             quota_5h_20x: Some(281.91),
             quota_7d_20x: Some(1691.46),
+            quota_updated_at: Some("2026-07-30T08:20:35+00:00".to_owned()),
+            quota_policy_5h: None,
+            quota_rows: Vec::new(),
+            rating_updated_at: Some("2026-08-02T12:42:22Z".to_owned()),
+            rating_models: vec![
+                crate::runtime::tools::CodexRatingMetric {
+                    label: "GPT-5.6 Sol ultra".to_owned(),
+                    average: Some(9.0),
+                    count: Some(27),
+                },
+                crate::runtime::tools::CodexRatingMetric {
+                    label: "DeepSeek V4 Flash max".to_owned(),
+                    average: Some(8.1),
+                    count: Some(57),
+                },
+            ],
             source_url: "https://codexradar.com/".to_owned(),
             feedback_url: "https://codexradar.com/".to_owned(),
         }),
