+++ qq-maid-core/src/runtime/tools/radar/tests/mod.rs
@@ -1,0 +1,2 @@
+mod integration;
+mod parse;
